feat:Add `standard_food_list` tibble
This commit adds the `standard_food_list` data frame to the `data.R` file. The tibble contains food names and codes of surveys supported by this package, and is read in from the "standard_food_list.csv" file using the `readr::read_csv()` function. Additionally, the commit adds documentation for the data frame using roxygen2-style comments. This change was made to provide internal functions with a standardized list of food names and codes for use in their analyses.