feat:Add `non_standard_food_list` tibble
This commit adds the `non_standard_food_list` tibble to the data.R file. The tibble contains non-standard food names and codes of surveys supported by this package, and is read in from the "non_standard_food_list.csv" file using the readr::read_csv() function. Additionally, the commit adds documentation for the data frame using roxygen2-style comments. This change was made to provide users and internal functions with a standardized matche of non-standard food names and codes for use in their analyses.