Update combined_food_list data
This commit updates the `combined_food_list` data to the latest version (v26072023). The data contains both standard and non-standard food names and codes of surveys supported by this package.

The update was performed by reading in the latest version of the CSV file using the `readr` package and the `here` package to locate the file in the data directory.

This commit also updates the documentation for the `combined_food_list` data to reflect the new version number.