Update DESCRIPTION file
This commit updates the `DESCRIPTION` file to include the new `stringdist` package as an import. The `stringdist` package is used in the `match_food_names` function to perform string distance matching between food names.

The update was performed by adding `stringdist` to the `Imports` field in the `DESCRIPTION` file.

This commit also updates the `Title` and `Description` field to include indicate that this package focuses on Nutrition data.