feat: Added `match_food_names_v2` function
Adds a function to match food names in a dataset to a standard list of food items based on the exact string match to an internal food list or a csv provided by the user.

The function provides diagnostics on the number of rows with no matches found and the list of foods with no matches found. The function is exported for use in other packages.